Move versioning stuff to separate files
