+++ src/plugins/intel_npu/src/plugin/include/compiled_model.hpp
@@ -13,65 +13,6 @@
 
 namespace intel_npu {
 
-typedef struct {
-    const char* major;
-    const char* minor;
-    const char* ovVersion;
-} Metadata;
-
-struct MetadataVersion {
-    uint32_t major;
-    uint32_t minor;
-} typedef MetadataVersion;
-
-struct OpenvinoVersion {
-    const std::string version;
-} typedef OpenvinoVersion;
-
-struct ModelLayout {
-    int something;
-    double somethingElse;
-} typedef ModelLayout;
-
-using metaIterator = std::vector<uint8_t>::iterator;
-
-// need constructors with minor and major fields
-struct Metadata_v1 {
-    MetadataVersion version;
-    OpenvinoVersion ovVersion;
-
-    static void version_check(std::vector<uint8_t>& blob, std::istream& stream);
-
-    std::vector<uint8_t> data();
-
-    void read_metadata(std::vector<uint8_t>::iterator& metadataIterator);
-
-    void write_metadata(std::ostream& stream);
-} typedef Metadata_v1;
-
-struct Metadata_v2 : Metadata_v1 {
-    ModelLayout layout;
-
-    std::vector<uint8_t> data();
-
-    void read_metadata(std::vector<uint8_t>::iterator& metadataIterator);
-
-    void write_metadata(std::ostream& stream);
-} typedef Metadata_v2;
-
-struct Metadata_v3 {
-    MetadataVersion version;
-    ModelLayout layout;
-    OpenvinoVersion ovVersion;
-    double extra;
-
-    std::vector<uint8_t> data();
-
-    void read_metadata(std::vector<uint8_t>::iterator metadataIterator);
-
-    void write_metadata(std::ostream& stream);
-} typedef Metadata_v3;
-
 class CompiledModel final : public ICompiledModel {
 public:
     /**
