+++ src/plugins/intel_npu/src/plugin/src/compiled_model.cpp
@@ -8,6 +8,7 @@
 #include <string_view>
 
 #include "async_infer_request.hpp"
+#include "model_version.hpp"
 #include "intel_npu/al/config/common.hpp"
 #include "intel_npu/al/config/compiler.hpp"
 #include "intel_npu/al/config/config.hpp"
@@ -138,187 +139,6 @@ std::shared_ptr<ov::ISyncInferRequest> CompiledModel::create_sync_infer_request(
         "the \"ov::ISyncInferRequest\" class");
 }
 
-std::vector<uint8_t> Metadata_v1::data() {
-    std::vector<uint8_t> metadata;
-    std::cout << "v1_data()\n";
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->version.major),
-    reinterpret_cast<uint8_t*>(&this->version.major) + sizeof(this->version.major));
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->version.minor),
-    reinterpret_cast<uint8_t*>(&this->version.minor) + sizeof(this->version.minor));
-
-    metadata.insert(metadata.end(), this->ovVersion.version.begin(), this->ovVersion.version.end());
-
-    return metadata;
-}
-
-// actually what should it return?
-void Metadata_v1::version_check(std::vector<uint8_t>& blob, std::istream& stream) {
-    constexpr std::string_view versionHeader{"OVNPU"}; // maybe put this some place else
-
-    size_t blobDataSize;
-    stream.read(reinterpret_cast<char*>(&blobDataSize), sizeof(size_t));
-    if (blobDataSize == blob.size() - sizeof(blobDataSize)) {
-        OPENVINO_THROW("Imported blob is not versioned");
-    }
-
-    auto metadataIterator {blob.begin() + blobDataSize};
-    char* blobVersionHeader = new char[versionHeader.size() + 1];
-    std::copy(metadataIterator, metadataIterator + versionHeader.size(), blobVersionHeader);
-    std::cout << "header: " << blobVersionHeader << '\n';
-
-    // should we consider the header name changes?
-    // if so, we might need multiple header #defines
-    if (versionHeader.compare(blobVersionHeader)) {
-        std::cout << "expected header: " << versionHeader << "\n actual header: " << blobVersionHeader << '\n';
-        OPENVINO_THROW("Version header mismatch or missing");
-    }
-    metadataIterator += versionHeader.size();
-
-    MetadataVersion metaVersion;
-    std::copy(metadataIterator, metadataIterator + sizeof(uint32_t), &metaVersion.major);
-    metadataIterator += sizeof(uint32_t);
-    std::cout << "major: " << metaVersion.major;
-
-    std::copy(metadataIterator, metadataIterator + sizeof(uint32_t), &metaVersion.minor);
-    std::cout << "\nminor: " << metaVersion.minor << '\n';
-    metadataIterator += sizeof(uint32_t);
-
-    std::vector<uint8_t> metaVec(metadataIterator, blob.end() - sizeof(size_t));
-    std::vector<uint8_t>::iterator metaIt = metaVec.begin();
-    // move this to another function?
-    if (metaVersion.major == 1) {
-        if (metaVersion.minor > 1 && metaVersion.minor < 5) {
-            std::cout << "We got Metadata_v1\n";
-            Metadata_v1 metav1;
-            metav1.version = metaVersion;
-            metav1.read_metadata(metaIt);
-        } else if (metaVersion.minor > 6) {
-            std::cout << "We got Metadata_v2\n";
-            Metadata_v2 metav2;
-            metav2.version = metaVersion;
-            metav2.read_metadata(metaIt);
-        }
-    } else if (metaVersion.major == 2) {
-        if (metaVersion.minor > 0) {
-            std::cout << "We got Metadata_v3\n";`
-            Metadata_v3 metav3;
-            metav3.version = metaVersion;
-            metav3.read_metadata(metaIt);
-        }
-    }
-}
-
-// should we check for header here or create a function called version_metadata_handler
-// which checks for header and calls read_metadata
-void Metadata_v1::read_metadata(std::vector<uint8_t>::iterator& metadataIterator) {
-    /*
-        is there a way needed to assert the version?
-        or do we still want to check for it?
-        after all, we can orchestrate everything using metadata major;minor
-    */
-    std::cout << "inside read: " << this->version.major << "." << this->version.minor << '\n';
-    ov::Version ourOvVersion = ov::get_openvino_version();
-    size_t ovVersionSize = strlen(ourOvVersion.buildNumber);
-    char* blobOvVersion = new char[ovVersionSize + 1];
-    std::copy(metadataIterator, metadataIterator + ovVersionSize, blobOvVersion);
-    std::cout << "blob ov version: " << blobOvVersion << '\n';
-    metadataIterator += ovVersionSize;
-}
-
-void Metadata_v1::write_metadata(std::ostream& stream) {
-    std::cout << "v1_write()\n";
-    const auto metav1_data = this->data();
-
-    stream.write(reinterpret_cast<const char*>(metav1_data.data()), metav1_data.size());
-}
-
-std::vector<uint8_t> Metadata_v2::data() {
-    std::cout << "v2_data()\n";
-    auto metadata {Metadata_v1::data()};
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->layout.something),
-    reinterpret_cast<uint8_t*>(&this->layout.something) + sizeof(this->layout.something));
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->layout.somethingElse),
-    reinterpret_cast<uint8_t*>(&this->layout.somethingElse) + sizeof(this->layout.somethingElse));
-
-    return metadata;
-}
-
-void Metadata_v2::read_metadata(std::vector<uint8_t>::iterator& metadataIterator) {
-    std::cout << "v2_read\n";
-    std::cout << "inside read: " << this->version.major << "." << this->version.minor << '\n';
-    Metadata_v1::read_metadata(metadataIterator);
-
-    memcpy(&this->layout.something, &(*metadataIterator), sizeof(this->layout.something));
-    std::cout << "something: " << this->layout.something << '\n';
-    metadataIterator += sizeof(int);
-
-    memcpy(&this->layout.somethingElse, &(*metadataIterator), sizeof(this->layout.somethingElse));
-    std::cout << "somethingElse: " << this->layout.somethingElse << '\n';
-    metadataIterator += sizeof(double);
-}
-
-void Metadata_v2::write_metadata(std::ostream& stream) {
-    std::cout << "v2_write()\n";
-    const auto metav2_data = this->data();
-    stream.write(reinterpret_cast<const char*>(metav2_data.data()), metav2_data.size());
-}
-
-std::vector<uint8_t> Metadata_v3::data() {
-    std::vector<uint8_t> metadata;
-    std::cout << "v1_data()\n";
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->version.major),
-    reinterpret_cast<uint8_t*>(&this->version.major) + sizeof(this->version.major));
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->version.minor),
-    reinterpret_cast<uint8_t*>(&this->version.minor) + sizeof(this->version.minor));
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->layout.something),
-    reinterpret_cast<uint8_t*>(&this->layout.something) + sizeof(this->layout.something));
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->layout.somethingElse),
-    reinterpret_cast<uint8_t*>(&this->layout.somethingElse) + sizeof(this->layout.somethingElse));
-
-    metadata.insert(metadata.end(), this->ovVersion.version.begin(), this->ovVersion.version.end());
-
-    metadata.insert(metadata.end(), reinterpret_cast<uint8_t*>(&this->extra),
-    reinterpret_cast<uint8_t*>(&this->extra) + sizeof(this->extra));
-
-    return metadata;
-}
-
-void Metadata_v3::read_metadata(std::vector<uint8_t>::iterator metadataIterator) {
-    std::cout << "inside read: " << this->version.major << "." << this->version.minor << '\n';
-
-    memcpy(&this->layout.something, &(*metadataIterator), sizeof(this->layout.something));
-    metadataIterator += sizeof(this->layout.something);
-    std::cout << "layout.something: " << this->layout.something << '\n';
-    memcpy(&this->layout.somethingElse, &(*metadataIterator), sizeof(this->layout.somethingElse));
-    metadataIterator += sizeof(this->layout.somethingElse);
-    std::cout << "layout.somethingElse: " << this->layout.somethingElse << '\n';
-
-    ov::Version ourOvVersion = ov::get_openvino_version();
-    size_t ovVersionSize = strlen(ourOvVersion.buildNumber);
-    char* blobOvVersion = new char[ovVersionSize + 1];
-    std::copy(metadataIterator, metadataIterator + ovVersionSize, blobOvVersion);
-    std::cout << "blob ov version: " << blobOvVersion << '\n';
-    metadataIterator += ovVersionSize;
-
-    memcpy(&this->extra, &(*metadataIterator), sizeof(this->extra));
-    metadataIterator += sizeof(this->extra);
-    std::cout << "extra: " << this->extra << '\n';
-}
-
-void Metadata_v3::write_metadata(std::ostream& stream) {
-    const auto metav3_data = this->data();
-
-    stream.write(reinterpret_cast<const char*>(metav3_data.data()), metav3_data.size());
-}
-
 void CompiledModel::export_model(std::ostream& stream) const {
     _logger.debug("CompiledModel::export_model");
     const auto&& blob = _compiler->getCompiledNetwork(_networkPtr);
@@ -328,7 +148,7 @@ void CompiledModel::export_model(std::ostream& stream) const {
     stream.write(metaHeader.data(), metaHeader.length());
     std::cout << "metaHeader size: " << metaHeader.length() << '\n';
 
-    MetadataVersion metaVersion = {2, 4};
+    MetadataVersion metaVersion = {1, 3};
     std::cout << "meta version sizes: " << metaVersion.major << " " << metaVersion.minor << '\n';
     OpenvinoVersion ovVersion = {"2024.5.0-16678-090da7b5376-blob_commit"};
     std::cout << "ovversion size: " << ovVersion.version.size() << '\n';
@@ -338,7 +158,7 @@ void CompiledModel::export_model(std::ostream& stream) const {
     Metadata_v2 metav2 = {metaVersion, ovVersion, layout};
     Metadata_v3 metav3 = {metaVersion, layout, ovVersion, 5.5};
 
-    metav3.write_metadata(stream);
+    metav1.write_metadata(stream);
 
     size_t blobSizeBeforeVersioning = blob.size();
     stream.write(reinterpret_cast<const char*>(&blobSizeBeforeVersioning), sizeof(size_t));
